(feat) lpss: 增加 OpenCV 图像转换支持（模拟 cv_bridge）

+++ samples/camera/mv/sample_mv_auto_calib.cpp
@@ -214,7 +214,7 @@ int main(int argc, char *argv[]) {
     int flags = 0;
     rm::MvCamera capture(rm::CameraConfig::create(rm::GrabMode::Continuous, rm::RetrieveMode::OpenCV));
 
-    int exposure{10000};
+    double exposure{10000};
     int gain{128};
     int r_gain{100};
     int g_gain{100};
